context & hooks folders for better organization
