+++ client/src/component/xterm/xterm.jsx
@@ -1,11 +1,9 @@
 import { useEffect, useRef } from "react"
 import { Terminal } from "@xterm/xterm"
-import "./xterm.css"
-import { useOptionsUpdate } from "../lab/lab-context.jsx"
-import { command_handler } from "./commands/handler.js"
-import { useTerminal } from "./terminal-context.jsx"
 import { FitAddon } from "@xterm/addon-fit"
-import { useSocketIO } from "../../util/socketio-context.jsx"
+import "./xterm.css"
+import { useOptionsUpdate, useSocketIO, useTerminal } from "../../hooks"
+import { command_handler } from "./commands/handler.js"
 
 const XTerm = ({ setShowFixMe }) => {
   const terminalRef = useRef(null)
