+++ client/src/hooks/socketio-hooks.jsx
@@ -1,0 +1,6 @@
+import { useContext } from 'react'
+import { SocketIOContext } from '../context'
+
+export const useSocketIO = () => {
+    return useContext(SocketIOContext)
+  }
