+++ client/src/hooks/terminal-hooks.jsx
@@ -1,0 +1,4 @@
+import { useContext } from "react"
+import { TerminalContext } from "../context"
+
+export const useTerminal = () => useContext(TerminalContext)
